Refactor: move out pullreq pipelines and shared code
